[S12P31A502-487] refactor: Modify running time

See merge request s12-final/S12P31A502!151

+++ backend/src/main/java/com/ssafy/yoittang/member/domain/dto/response/MemberProfileResponse.java
@@ -16,7 +16,7 @@ public record MemberProfileResponse(
         Integer followerCount,
         Boolean isFollow,
         RunningTimeResponse time,
-        Integer totalDistance,
+        Double totalDistance,
         Integer totalTileCount,
         List<CourseSummaryResponse> courses
 ) {
